A2G-1-Refactoring
