+++ pom.xml
@@ -62,17 +62,12 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
 		</dependency>
-		
-
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-devtools</artifactId>
 			<scope>runtime</scope>
 			<optional>true</optional>
 		</dependency>
-		
-	
-		
 		<dependency>
 			<groupId>org.projectlombok</groupId>
 			<artifactId>lombok</artifactId>
@@ -100,7 +95,6 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-websocket</artifactId>
 		</dependency>
-		
 		<dependency>
 		  <groupId>org.webjars</groupId>
 		  <artifactId>webjars-locator-core</artifactId>
@@ -124,6 +118,11 @@
 		  <groupId>org.webjars</groupId>
 		  <artifactId>jquery</artifactId>
 		  <version>3.1.1-1</version>
+		</dependency>
+		<dependency>
+			<groupId>org.apache.logging.log4j</groupId>
+			<artifactId>log4j</artifactId>
+			<version>2.17.1</version>
 		</dependency>
 	</dependencies>
