--- src/main/java/com/whiteroom/ani2gether/WebMvcConfig.java
@@ -1,39 +1,0 @@
-package com.whiteroom.ani2gether;
-
-import java.util.Locale;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.LocaleResolver;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.springframework.web.servlet.i18n.LocaleChangeInterceptor;
-import org.springframework.web.servlet.i18n.SessionLocaleResolver;
-
-@Configuration
-public class WebMvcConfig implements WebMvcConfigurer{
-	
-	
-	// following code adds localization (i18n) support
-	// by default localization files are added in /resources/ folder 
-	// with `messages_XX.properties` filename 
-	// check spring.messages.basename in application.properties to know which basepath is used
-	@Bean(name = "localeResolver")
-	  public LocaleResolver localeResolver() {
-	    SessionLocaleResolver slr = new SessionLocaleResolver();
-	    slr.setDefaultLocale(Locale.getDefault());
-	    return slr;
-	  }
-	
-	@Bean
-	public LocaleChangeInterceptor localeChangeInterceptor() {
-	    LocaleChangeInterceptor lci = new LocaleChangeInterceptor();
-	    lci.setParamName("lang");
-	    return lci;
-	}
-
-	@Override
-	public void addInterceptors(InterceptorRegistry registry) {
-	    registry.addInterceptor(localeChangeInterceptor());
-	}
-}
