--- src/main/java/com/whiteroom/ani2gether/WebSocketConfig.java
@@ -1,29 +1,0 @@
-package com.whiteroom.ani2gether;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
-import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
-import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
-
-@Configuration
-@EnableWebSocketMessageBroker
-public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
-
-  @Override
-  public void configureMessageBroker(MessageBrokerRegistry config) {
-    config.enableSimpleBroker("/topic");
-    config.setApplicationDestinationPrefixes("/rpg");
-  }
-
-  @Override
-  public void registerStompEndpoints(StompEndpointRegistry registry) {
-    //registry.addEndpoint("/gs-guide-websocket").withSockJS();
-    registry.addEndpoint("/mob").withSockJS();
-    
-  }
-  
-  
-  
-
-}
