+++ src/main/java/com/whiteroom/ani2gether/domain/model/UserCommand.java
@@ -1,0 +1,19 @@
+package com.whiteroom.ani2gether.domain.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserCommand {
+
+	private String command;
+	private double timestamp;
+
+}
