+++ src/main/resources/application.properties
@@ -1,11 +1,10 @@
 server.port=8080
-server.ip=192.168.100.200
+server.ip=192.168.0.7
 
 #
 # i18n properties
 #
 spring.messages.basename=localization/test/messages, localization/index/messages
 
-# a gde ugol v moodle
